--- src/main/java/com/tn/server/common/response/ApiResponse.java
@@ -1,24 +1,0 @@
-package com.tn.server.common.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ApiResponse<T> {
-    private String code;
-    private String message;
-    private T data;
-
-    public static <T> ApiResponse<T> success(String message, T data) {
-        return ApiResponse.<T>builder()
-                .code("SUCCESS")
-                .message(message)
-                .data(data)
-                .build();
-    }
-
-    public static <T> ApiResponse<T> success(T data) {
-        return success("요청이 성공적으로 처리되었습니다.", data);
-    }
-}
