--- src/main/java/com/tn/server/common/response/ErrorResponse.java
@@ -1,11 +1,0 @@
-package com.tn.server.common.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ErrorResponse {
-    private final String code;
-    private final String message;
-}
