+++ src/main/java/com/tn/server/controller/ImageTestController.java
@@ -1,6 +1,6 @@
 package com.tn.server.controller;
 
-import com.tn.server.common.response.ApiResponse;
+import com.tn.server.dto.common.ApiResponse;
 import com.tn.server.service.image.ImageManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Profile;
@@ -25,7 +25,7 @@ public class ImageTestController {
         // 테스트용이므로 경로는 "tests"로 고정
         String result = imageManager.upload(file, "tests", isSecret);
 
-        return ResponseEntity.ok(ApiResponse.success(
+        return ResponseEntity.ok(ApiResponse.success("업로드에 성공했습니다.",
                 (isSecret ? "[비공개-키 반환] " : "[공개-URL 반환] ") + result
         ));
     }
@@ -37,7 +37,7 @@ public class ImageTestController {
             @RequestParam(value = "isSecret", defaultValue = "false") boolean isSecret
     ) {
         imageManager.delete(keyOrUrl, isSecret);
-        return ResponseEntity.ok(ApiResponse.success("삭제 요청 완료 (R2 확인 필요)"));
+        return ResponseEntity.ok(ApiResponse.success("삭제 요청 완료","삭제 요청 완료 (R2 확인 필요)"));
     }
 
     // 3. 다운로드용 Presigned URL 발급 테스트 (GET)
@@ -46,7 +46,7 @@ public class ImageTestController {
             @RequestParam("key") String key
     ) {
         String downloadUrl = imageManager.getPresignedGetUrl(key);
-        return ResponseEntity.ok(ApiResponse.success(downloadUrl));
+        return ResponseEntity.ok(ApiResponse.success("URL 발급에 성공했습니다.", downloadUrl));
     }
 
     // 4. URL에서 이미지 다운로드 후 업로드 테스트
@@ -57,7 +57,7 @@ public class ImageTestController {
     ) {
         String result = imageManager.uploadFromUrl(imageUrl, "test-url-uploads", isSecret);
         return ResponseEntity.ok(ApiResponse.success(
-                (isSecret ? "[비공개-키 반환] " : "[공개-URL 반환] ") + result
+                "URL 발급에 성공했습니다.",(isSecret ? "[비공개-키 반환] " : "[공개-URL 반환] ") + result
         ));
     }
 }
